pyproject.toml
[tool.poetry]
name = "fastapi_todo"
version = "0.1"
description = "fastapi todo app"
authors = ["hamedsh <author@email.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
fastapi = "^0.68.1"
SQLAlchemy = "^1.4.23"
mypy = "^0.910"
uvicorn = "^0.15.0"
sqlalchemy-stubs = "^0.4"
psycopg2-binary = "^2.9.1"
#psycopg2 = "^2.9.1"
pydantic = {extras = ["email"], version = "^1.8.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
nitpick = "^0.27.0"
wemake-python-styleguide = "^0.15.3"
mypy = "^0.910"
tox = "^3.24.3"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
alembic = "^1.7.1"

[tool.poetry.extras]
lint = ["pylint"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nitpick]
style = "https://raw.githubusercontent.com/hamedsh/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
